Expand on '.editor is active' line in readme
This really confused me at first and I think it will continue to confuse people. I had to do some digging through issues to find this. It might be a bit of a ugly addition to the docs but saves users of this package the headache of digging through closed issues =D